Move utils into algs library
